optimize create function and procedure validator
